refactor: QK config object
