+++ QueryKit/Configuration/QueryKitSettings.cs
@@ -1,0 +1,31 @@
+namespace QueryKit.Configuration;
+
+using System.Linq.Expressions;
+using QueryKit.Operators;
+
+public class QueryKitSettings
+{
+    public QueryKitPropertyMappings PropertyMappings { get; set; } = new QueryKitPropertyMappings();
+    public string EqualsOperator { get; set; } = ComparisonOperator.EqualsOperator().Operator();
+    public string NotEqualsOperator { get; set; } = ComparisonOperator.NotEqualsOperator().Operator();
+    public string GreaterThanOperator { get; set; } = ComparisonOperator.GreaterThanOperator().Operator();
+    public string LessThanOperator { get; set; } = ComparisonOperator.LessThanOperator().Operator();
+    public string GreaterThanOrEqualOperator { get; set; } = ComparisonOperator.GreaterThanOrEqualOperator().Operator();
+    public string LessThanOrEqualOperator { get; set; } = ComparisonOperator.LessThanOrEqualOperator().Operator();
+    public string ContainsOperator { get; set; } = ComparisonOperator.ContainsOperator().Operator();
+    public string StartsWithOperator { get; set; } = ComparisonOperator.StartsWithOperator().Operator();
+    public string EndsWithOperator { get; set; } = ComparisonOperator.EndsWithOperator().Operator();
+    public string NotContainsOperator { get; set; } = ComparisonOperator.NotContainsOperator().Operator();
+    public string NotStartsWithOperator { get; set; } = ComparisonOperator.NotStartsWithOperator().Operator();
+    public string NotEndsWithOperator { get; set; } = ComparisonOperator.NotEndsWithOperator().Operator();
+    public string InOperator { get; set; } = ComparisonOperator.InOperator().Operator();
+    public string AndOperator { get; set; } = LogicalOperator.AndOperator.Operator();
+    public string OrOperator { get; set; } = LogicalOperator.OrOperator.Operator();
+    public string CaseInsensitiveAppendix { get; set; } = "*";
+    public bool AllowUnknownProperties { get; set; }
+    
+    public QueryKitPropertyMapping<TModel> Property<TModel>(Expression<Func<TModel, object>>? propertySelector)
+    {
+        return PropertyMappings.Property(propertySelector);
+    }
+}
