+++ QueryKit/SortParser.cs
@@ -2,6 +2,7 @@ namespace QueryKit;
 
 using System.Linq.Expressions;
 using System.Reflection;
+using Configuration;
 using Exceptions;
 using Sprache;
